extracted inner classes, fixed tests
