+++ v2/googlecloud-to-elasticsearch/src/main/java/com/google/cloud/teleport/v2/elasticsearch/options/package-info.java
@@ -17,4 +17,4 @@
 /**
  * Package info for elasticsearch options.
  */
-package com.google.cloud.teleport.v2.elasticsearch.templates;
+package com.google.cloud.teleport.v2.elasticsearch.options;
