+++ v2/googlecloud-to-elasticsearch/src/main/resources/csv-to-elasticsearch-command-spec.json
@@ -1,0 +1,8 @@
+{
+  "mainClass": "com.google.cloud.teleport.v2.elasticsearch.templates.CsvToElasticsearch",
+  "classPath": "/template/googlecloud-to-elasticsearch/*:/template/googlecloud-to-elasticsearch/libs/*:/template/googlecloud-to-elasticsearch/classes",
+  "defaultParameterValues": {
+    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\", \"goog-dataflow-provided-template-name\":\"csv_to_elasticsearch\"}"
+  }
+}
+
