--- v2/googlecloud-to-elasticsearch/src/main/resources/googlecloud-to-elasticsearch-command-spec.json
@@ -1,7 +1,0 @@
-{
-  "mainClass": "com.google.cloud.teleport.v2.elasticsearch.templates.GoogleCloudToElasticsearchMain",
-  "classPath": "/template/googlecloud-to-elasticsearch/*:/template/googlecloud-to-elasticsearch/libs/*:/template/googlecloud-to-elasticsearch/classes",
-  "defaultParameterValues": {
-    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\", \"goog-dataflow-provided-template-name\":\"googlecloud_to_elasticsearch\"}"
-}
-}
